+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/ObjectWaitBlocked.kt
@@ -8,7 +8,8 @@ class ObjectWaitBlocked(val objectContext: ObjectNotifyContext, timed: Boolean) 
     BlockedOperation(
         timed,
         ResourceInfo(
-            System.identityHashCode(objectContext.objReference.get()), ResourceType.CONDITION)) {
+            System.identityHashCode(objectContext.lockContext.lockReference.get()),
+            ResourceType.CONDITION)) {
   override fun unblockThread(tid: Long, type: InterruptionType): Any? {
     if (objectContext.unblockThread(tid, type)) {
       return objectContext.getSyncObject()