+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/primitives/ObjectNotifyContext.kt
@@ -1,6 +1,5 @@
 package org.pastalab.fray.core.concurrency.primitives
 
-import java.lang.ref.WeakReference
 import org.pastalab.fray.core.ThreadContext
 import org.pastalab.fray.core.concurrency.operations.InterruptionType
 import org.pastalab.fray.core.concurrency.operations.ObjectWaitBlocked
@@ -8,10 +7,10 @@ import org.pastalab.fray.core.concurrency.operations.ObjectWakeBlocked
 import org.pastalab.fray.rmi.ThreadState
 
 class ObjectNotifyContext(lockContext: LockContext, obj: Any) : SignalContext(lockContext) {
-  val objReference = WeakReference(obj)
 
   override fun sendSignalToObject() {
-    objReference.get()?.let { obj -> synchronized(obj) { (obj as Object).notifyAll() } }
+    val obj = lockContext.lockReference.get() ?: return
+    synchronized(obj) { (obj as Object).notifyAll() }
   }
 
   override fun updatedThreadContextDueToUnblock(